requirements.txt
asgiref==3.7.2
Babel==2.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.2
djangorestframework==3.14.0
ghp-import==2.1.0
idna==3.6
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.5
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-material==9.5.11
mkdocs-material-extensions==1.3.1
packaging==23.2
paginate==0.5.6
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
Pygments==2.17.2
pymdown-extensions==10.7
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
